requirements.txt
selenium==4.9.0
boto3==1.26.142
pandas==1.5.3
PyYAML==6.0.1